Update AIBot jar in demo
