+++ tmail-backend/tmail-third-party/ai-bot/AIBotMailet.adoc
@@ -1,0 +1,24 @@
+= AIBotMailet
+
+`AIBotMailet` is a mailet designed to automatically reply to emails sent to a specific recipient (e.g., gpt@linagora.com). It uses the OpenAI API by default but can be configured to use any LLM API supported by https://github.com/langchain4j/[langchain4j].
+
+== Configuration Example
+
+[source,xml]
+----
+<mailet match="com.linagora.tmail.mailet.RecipientsContain=gpt@tmail.com" class="com.linagora.tmail.mailet.AIBotMailet">
+      <apiKey>{put you api key here}</apiKey>
+      <botAddress>gpt@tmail.com</botAddress>
+      <model>linagoras-chitchat</model>
+      <baseURL>https://ai.linagora.com/api</baseURL>
+</mailet>
+----
+
+=== Parameters
+
+The mailet configuration requires the following parameters:
+
+* `apiKey`: The API key for accessing the LLM API service.
+* `botAddress`: The email address used to send replies generated by the LLM model.
+* `baseURL`: The URL of an LLM API compatible with OpenAI .e.g. https://ai.linagora.com/api. Optional. Defaults to the official OpenAI API base url.
+* `model`: The OpenAI LLM model to be used. Optional. Defaults to `gpt-3.5-turbo` by langchain4j. You can use the `gpt-4o-mini` model and `demo` API key for testing purpose.
