+++ tmail-backend/tmail-third-party/ai-bot/src/main/java/com/linagora/tmail/mailet/LlmModel.java
@@ -1,0 +1,4 @@
+package com.linagora.tmail.mailet;
+
+public record LlmModel(String modelName) {
+}
