--- tmail-backend/tmail-third-party/open-ai/RecipientsContain.adoc
@@ -1,13 +1,0 @@
-=== RecipientsContain
-
-`RecipientsContain` is a matcher that checks if the recipients of an email include all the addresses specified in the match condition.
-
-This matcher will return all recipients of the email if all specified addresses are present among the recipients. Otherwise, it returns an empty list.
-
-== Configuration Example
-
-[source,xml]
-----
-<mailet match="com.linagora.tmail.mailet.RecipientsContain=gpt1@linagora.com,gpt2@linagora.com" class="org.apache.james.transport.mailets.Null">
-</mailet>
-----
