--- tmail-backend/tmail-third-party/open-ai/src/main/java/com/linagora/tmail/mailet/LlmModel.java
@@ -1,4 +1,0 @@
-package com.linagora.tmail.mailet;
-
-public record LlmModel(String modelName) {
-}
